Changed the amount of visible slides
calculated the amount of visible slides based on a newly added `desiredWidth` value. Changed `getWidth()` for this calculation. also updated `enableInfiniteMode()` to take into account, that there are more slides to be cloned.  (based on the slick calculation). Everything ist completely responsive (mobile first) designed.

Furthermore: changd the scss a little to get it to work better with bootstrap.